fixes for migration to work
