+++ hedera-node/hedera-app-spi/src/main/java/com/hedera/node/app/spi/Service.java
@@ -29,9 +29,6 @@ import java.util.Set;
  * crypto-service, token-service etc.,
  */
 public interface Service {
-    SemanticVersion RELEASE_045_VERSION = new SemanticVersion(0, 45, 0, "SNAPSHOT", "");
-    SemanticVersion RELEASE_MIGRATION_VERSION = new SemanticVersion(0, 45, 1, "SNAPSHOT", "");
-
     /**
      * Returns the name of the service. This name must be unique for each service deployed on the
      * application.
