+++ hedera-node/hedera-app/src/test/java/com/hedera/node/app/AppTestBase.java
@@ -16,7 +16,7 @@
 
 package com.hedera.node.app;
 
-import static com.hedera.node.app.spi.Service.RELEASE_045_VERSION;
+import static com.hedera.node.app.spi.fixtures.state.TestSchema.CURRENT_VERSION;
 import static java.util.Objects.requireNonNull;
 
 import com.hedera.hapi.node.base.AccountID;
@@ -237,8 +237,8 @@ public class AppTestBase extends TestBase implements TransactionFactory, Scenari
     }
 
     public static final class TestAppBuilder {
-        private SemanticVersion softwareVersion = RELEASE_045_VERSION;
-        private SemanticVersion hapiVersion = RELEASE_045_VERSION;
+        private SemanticVersion softwareVersion = CURRENT_VERSION;
+        private SemanticVersion hapiVersion = CURRENT_VERSION;
         private Set<Service> services = new LinkedHashSet<>();
         private TestConfigBuilder configBuilder = HederaTestConfigBuilder.create();
         private NodeInfo selfNodeInfo = null;
