+++ hedera-node/hedera-app/src/test/java/com/hedera/node/app/state/recordcache/RecordCacheServiceTest.java
@@ -16,7 +16,7 @@
 
 package com.hedera.node.app.state.recordcache;
 
-import static com.hedera.node.app.spi.Service.RELEASE_045_VERSION;
+import static com.hedera.node.app.spi.fixtures.state.TestSchema.CURRENT_VERSION;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.verify;
 
@@ -43,10 +43,10 @@ final class RecordCacheServiceTest {
     @Test
     void schema(@Mock final SchemaRegistry registry) {
         final var svc = new RecordCacheService();
-        svc.registerSchemas(registry, RELEASE_045_VERSION);
+        svc.registerSchemas(registry, CURRENT_VERSION);
         verify(registry).register(captor.capture());
         final var schema = captor.getValue();
-        assertThat(schema.getVersion()).isEqualTo(RELEASE_045_VERSION);
+        assertThat(schema.getVersion()).isEqualTo(CURRENT_VERSION);
         assertThat(schema.statesToCreate()).hasSize(1);
     }
 }
