+++ hedera-node/hedera-consensus-service-impl/src/main/java/com/hedera/node/app/service/consensus/impl/ConsensusServiceImpl.java
@@ -17,17 +17,12 @@
 package com.hedera.node.app.service.consensus.impl;
 
 import com.hedera.hapi.node.base.SemanticVersion;
-import com.hedera.hapi.node.base.TopicID;
-import com.hedera.hapi.node.state.consensus.Topic;
 import com.hedera.node.app.service.consensus.ConsensusService;
-import com.hedera.node.app.service.consensus.impl.codecs.ConsensusServiceStateTranslator;
 import com.hedera.node.app.service.consensus.impl.schemas.InitialModServiceConsensusSchema;
 import com.hedera.node.app.service.mono.state.merkle.MerkleTopic;
 import com.hedera.node.app.service.mono.utils.EntityNum;
-import com.hedera.node.app.spi.state.MigrationContext;
 import com.hedera.node.app.spi.state.Schema;
 import com.hedera.node.app.spi.state.SchemaRegistry;
-import com.hedera.node.app.spi.state.WritableKVStateBase;
 import com.swirlds.merkle.map.MerkleMap;
 import edu.umd.cs.findbugs.annotations.NonNull;
 
@@ -44,25 +39,7 @@ public final class ConsensusServiceImpl implements ConsensusService {
     @Override
     public void registerSchemas(@NonNull SchemaRegistry registry, final SemanticVersion version) {
         // We intentionally ignore the given (i.e. passed-in) version in this method
-        registry.register(consensusSchema(RELEASE_045_VERSION));
-
-        registry.register(new Schema(RELEASE_MIGRATION_VERSION) {
-
-            @Override
-            public void migrate(@NonNull MigrationContext ctx) {
-                if (fs != null) {
-                    System.out.println("BBM: running consensus migration...");
-
-                    var ts = ctx.newStates().<TopicID, Topic>get(TOPICS_KEY);
-                    ConsensusServiceStateTranslator.migrateFromMerkleToPbj(fs, ts);
-                    if (ts.isModified()) ((WritableKVStateBase) ts).commit();
-
-                    fs = null;
-
-                    System.out.println("BBM: finished consensus");
-                }
-            }
-        });
+        registry.register(consensusSchema(version));
     }
 
     public void setFromState(MerkleMap<EntityNum, MerkleTopic> fs) {
@@ -70,6 +47,6 @@ public final class ConsensusServiceImpl implements ConsensusService {
     }
 
     private Schema consensusSchema(final SemanticVersion version) {
-        return new InitialModServiceConsensusSchema(version);
+        return new InitialModServiceConsensusSchema(version, fs);
     }
 }
