+++ hedera-node/hedera-consensus-service-impl/src/main/java/com/hedera/node/app/service/consensus/impl/schemas/InitialModServiceConsensusSchema.java
@@ -16,12 +16,19 @@
 
 package com.hedera.node.app.service.consensus.impl.schemas;
 
+import static com.hedera.node.app.service.consensus.impl.ConsensusServiceImpl.TOPICS_KEY;
+
 import com.hedera.hapi.node.base.SemanticVersion;
 import com.hedera.hapi.node.base.TopicID;
 import com.hedera.hapi.node.state.consensus.Topic;
-import com.hedera.node.app.service.consensus.impl.ConsensusServiceImpl;
+import com.hedera.node.app.service.consensus.impl.codecs.ConsensusServiceStateTranslator;
+import com.hedera.node.app.service.mono.state.merkle.MerkleTopic;
+import com.hedera.node.app.service.mono.utils.EntityNum;
+import com.hedera.node.app.spi.state.MigrationContext;
 import com.hedera.node.app.spi.state.Schema;
 import com.hedera.node.app.spi.state.StateDefinition;
+import com.hedera.node.app.spi.state.WritableKVStateBase;
+import com.swirlds.merkle.map.MerkleMap;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Set;
 
@@ -32,13 +39,31 @@ import java.util.Set;
  * this schema is always correct for the current version of the software.
  */
 public class InitialModServiceConsensusSchema extends Schema {
-    public InitialModServiceConsensusSchema(SemanticVersion version) {
+    private MerkleMap<EntityNum, MerkleTopic> fs;
+
+    public InitialModServiceConsensusSchema(SemanticVersion version, final MerkleMap<EntityNum, MerkleTopic> fs) {
         super(version);
+        this.fs = fs;
     }
 
     @NonNull
     @Override
     public Set<StateDefinition> statesToCreate() {
-        return Set.of(StateDefinition.inMemory(ConsensusServiceImpl.TOPICS_KEY, TopicID.PROTOBUF, Topic.PROTOBUF));
+        return Set.of(StateDefinition.inMemory(TOPICS_KEY, TopicID.PROTOBUF, Topic.PROTOBUF));
+    }
+
+    @Override
+    public void migrate(@NonNull MigrationContext ctx) {
+        if (fs != null) {
+            System.out.println("BBM: running consensus migration...");
+
+            var ts = ctx.newStates().<TopicID, Topic>get(TOPICS_KEY);
+            ConsensusServiceStateTranslator.migrateFromMerkleToPbj(fs, ts);
+            if (ts.isModified()) ((WritableKVStateBase) ts).commit();
+
+            fs = null;
+
+            System.out.println("BBM: finished consensus");
+        }
     }
 }
