+++ hedera-node/hedera-file-service-impl/src/main/java/com/hedera/node/app/service/file/impl/FileServiceImpl.java
@@ -16,11 +16,8 @@
 
 package com.hedera.node.app.service.file.impl;
 
-import com.hedera.hapi.node.base.FileID;
 import com.hedera.hapi.node.base.SemanticVersion;
-import com.hedera.hapi.node.state.file.File;
 import com.hedera.node.app.service.file.FileService;
-import com.hedera.node.app.service.file.impl.codec.FileServiceStateTranslator;
 import com.hedera.node.app.service.file.impl.schemas.InitialModFileGenesisSchema;
 import com.hedera.node.app.service.mono.files.DataMapFactory;
 import com.hedera.node.app.service.mono.files.HFileMeta;
@@ -29,15 +26,9 @@ import com.hedera.node.app.service.mono.files.store.FcBlobsBytesStore;
 import com.hedera.node.app.service.mono.state.adapters.VirtualMapLike;
 import com.hedera.node.app.service.mono.state.virtual.VirtualBlobKey;
 import com.hedera.node.app.service.mono.state.virtual.VirtualBlobValue;
-import com.hedera.node.app.spi.state.MigrationContext;
-import com.hedera.node.app.spi.state.Schema;
 import com.hedera.node.app.spi.state.SchemaRegistry;
-import com.hedera.node.app.spi.state.WritableKVStateBase;
 import com.hedera.node.config.ConfigProvider;
-import com.swirlds.common.threading.manager.AdHocThreadManager;
 import edu.umd.cs.findbugs.annotations.NonNull;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
 import javax.inject.Inject;
@@ -67,60 +58,6 @@ public final class FileServiceImpl implements FileService {
 
     @Override
     public void registerSchemas(@NonNull final SchemaRegistry registry, final SemanticVersion version) {
-        registry.register(new InitialModFileGenesisSchema(RELEASE_045_VERSION, configProvider));
-
-        registry.register(new Schema(RELEASE_MIGRATION_VERSION) {
-            @Override
-            public void migrate(@NonNull MigrationContext ctx) {
-                if (fss.get() != null) {
-                    var ts = ctx.newStates().<FileID, File>get(BLOBS_KEY);
-
-                    System.out.println("BBM: running file migration...");
-                    var allFileIds = extractFileIds(fss.get());
-                    var migratedFileIds = new ArrayList<Long>();
-                    allFileIds.forEach(fromFileIdRaw -> {
-                        var fromFileId = com.hederahashgraph.api.proto.java.FileID.newBuilder()
-                                .setFileNum(fromFileIdRaw)
-                                .build();
-                        var fromFileMeta = fileAttrs.get(fromFileId);
-                        // Note: if the file meta is null, then this file is more specialized
-                        // (e.g. contract bytecode) and will be migrated elsewhere
-                        if (fromFileMeta != null) {
-                            File toFile = FileServiceStateTranslator.stateToPbj(
-                                    fileContents.get(fromFileId), fromFileMeta, fromFileId);
-                            ts.put(
-                                    FileID.newBuilder()
-                                            .fileNum(fromFileId.getFileNum())
-                                            .build(),
-                                    toFile);
-                            migratedFileIds.add(fromFileIdRaw);
-                        }
-                    });
-
-                    if (ts.isModified()) ((WritableKVStateBase) ts).commit();
-
-                    fss = null;
-                    fileContents = null;
-                    fileAttrs = null;
-
-                    System.out.println("BBM:finished file migration. Migrated files: " + migratedFileIds);
-                }
-            }
-        });
-    }
-
-    private List<Long> extractFileIds(VirtualMapLike<VirtualBlobKey, VirtualBlobValue> fileStorage) {
-        final var fileIds = new ArrayList<Long>();
-        try {
-            fileStorage.extractVirtualMapData(
-                    AdHocThreadManager.getStaticThreadManager(),
-                    entry -> {
-                        fileIds.add((long) entry.left().getEntityNumCode());
-                    },
-                    1);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-        return fileIds;
+        registry.register(new InitialModFileGenesisSchema(version, configProvider, fss, fileContents, fileAttrs));
     }
 }
