+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/node/app/service/mono/ServicesState.java
@@ -115,7 +115,7 @@ public class ServicesState extends PartialNaryMerkleInternal
         implements MerkleInternal, SwirldState, StateChildrenProvider {
     private static final Logger log = LogManager.getLogger(ServicesState.class);
 
-    private static final long RUNTIME_CONSTRUCTABLE_ID = 0x8e300b0dfdafbb1aL;
+    private static final long RUNTIME_CONSTRUCTABLE_ID = 0x8e300b0dfdafbb10L;
     public static final ImmutableHash EMPTY_HASH = new ImmutableHash(new byte[DigestType.SHA_384.digestLength()]);
 
     // Only over-written when Platform deserializes a legacy version of the state
