+++ hedera-node/hedera-smart-contract-service-impl/src/main/java/com/hedera/node/app/service/contract/impl/state/InitialModServiceContractSchema.java
@@ -21,10 +21,18 @@ import com.hedera.hapi.node.state.common.EntityNumber;
 import com.hedera.hapi.node.state.contract.Bytecode;
 import com.hedera.hapi.node.state.contract.SlotKey;
 import com.hedera.hapi.node.state.contract.SlotValue;
+import com.hedera.node.app.service.contract.impl.ContractServiceImpl;
+import com.hedera.node.app.service.mono.state.migration.ContractStateMigrator;
+import com.hedera.node.app.service.mono.state.virtual.VirtualBlobKey;
 import com.hedera.node.app.spi.state.MigrationContext;
 import com.hedera.node.app.spi.state.Schema;
 import com.hedera.node.app.spi.state.StateDefinition;
+import com.hedera.node.app.spi.state.WritableKVState;
+import com.hedera.node.app.spi.state.WritableKVStateBase;
+import com.hedera.pbj.runtime.io.buffer.Bytes;
+import com.swirlds.common.threading.manager.AdHocThreadManager;
 import edu.umd.cs.findbugs.annotations.NonNull;
+import java.util.ArrayList;
 import java.util.Set;
 
 /**
@@ -45,7 +53,67 @@ public class InitialModServiceContractSchema extends Schema {
 
     @Override
     public void migrate(@NonNull final MigrationContext ctx) {
-        super.migrate(ctx);
+        if (ContractServiceImpl.getFromState() != null) {
+            System.out.println("BBM: migrating contract service");
+
+            System.out.println("BBM: migrating contract k/v storage...");
+            var result = ContractStateMigrator.migrateFromContractStorageVirtualMap(
+                    ContractServiceImpl.getFromState(),
+                    ContractServiceImpl.getToState(),
+                    ContractServiceImpl.getFlusher());
+            ContractServiceImpl.setFromState(null);
+            ContractServiceImpl.setToState(null);
+            System.out.println("BBM: finished migrating contract storage. Result: " + result);
+
+            System.out.println("BBM: migrating contract bytecode...");
+            WritableKVState<EntityNumber, Bytecode> bytecodeTs =
+                    ctx.newStates().get(InitialModServiceContractSchema.BYTECODE_KEY);
+            var migratedContractNums = new ArrayList<Integer>();
+            try {
+                ContractServiceImpl.getFss()
+                        .get()
+                        .extractVirtualMapData(
+                                AdHocThreadManager.getStaticThreadManager(),
+                                entry -> {
+                                    if (VirtualBlobKey.Type.CONTRACT_BYTECODE
+                                            == entry.left().getType()) {
+                                        var contractId = entry.left().getEntityNumCode();
+                                        var contents = entry.right().getData();
+                                        Bytes wrappedContents;
+                                        if (contents == null || contents.length < 1) {
+                                            System.out.println(
+                                                    "BBM: contract contents null for contractId " + contractId);
+                                            wrappedContents = Bytes.EMPTY;
+                                        } else {
+                                            System.out.println("BBM: migrating contract contents (length "
+                                                    + contents.length
+                                                    + ") for contractId "
+                                                    + contractId);
+                                            wrappedContents = Bytes.wrap(contents);
+                                        }
+                                        bytecodeTs.put(
+                                                EntityNumber.newBuilder()
+                                                        .number(contractId)
+                                                        .build(),
+                                                Bytecode.newBuilder()
+                                                        .code(wrappedContents)
+                                                        .build());
+                                        migratedContractNums.add(contractId);
+                                    }
+                                },
+                                1);
+            } catch (InterruptedException e) {
+                throw new RuntimeException(e);
+            }
+
+            System.out.println("BBM: finished migrating contract bytecode. Contract nums: " + migratedContractNums);
+
+            if (bytecodeTs.isModified()) ((WritableKVStateBase) bytecodeTs).commit();
+
+            ContractServiceImpl.setFileFs(null);
+
+            System.out.println("BBM: contract migration finished");
+        }
     }
 
     @NonNull
