+++ hedera-node/hedera-smart-contract-service-impl/src/test/java/com/hedera/node/app/service/contract/impl/test/ContractServiceImplTest.java
@@ -17,7 +17,7 @@
 package com.hedera.node.app.service.contract.impl.test;
 
 import static com.hedera.node.app.service.contract.impl.ContractServiceImpl.CONTRACT_SERVICE;
-import static com.hedera.node.app.spi.Service.RELEASE_045_VERSION;
+import static com.hedera.node.app.spi.fixtures.state.TestSchema.CURRENT_VERSION;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.mock;
@@ -39,7 +39,7 @@ class ContractServiceImplTest {
     void registersContractSchema() {
         final var captor = ArgumentCaptor.forClass(Schema.class);
         final var mockRegistry = mock(SchemaRegistry.class);
-        CONTRACT_SERVICE.registerSchemas(mockRegistry, RELEASE_045_VERSION);
+        CONTRACT_SERVICE.registerSchemas(mockRegistry, CURRENT_VERSION);
         verify(mockRegistry).register(captor.capture());
         assertInstanceOf(InitialModServiceContractSchema.class, captor.getValue());
     }
