+++ hedera-node/hedera-token-service-impl/src/main/java/com/hedera/node/app/service/token/impl/TokenServiceImpl.java
@@ -18,30 +18,11 @@ package com.hedera.node.app.service.token.impl;
 
 import static java.util.Objects.requireNonNull;
 
-import com.hedera.hapi.node.base.AccountID;
-import com.hedera.hapi.node.base.NftID;
 import com.hedera.hapi.node.base.SemanticVersion;
-import com.hedera.hapi.node.base.TokenID;
-import com.hedera.hapi.node.state.common.EntityIDPair;
-import com.hedera.hapi.node.state.common.EntityNumber;
 import com.hedera.hapi.node.state.token.Account;
-import com.hedera.hapi.node.state.token.NetworkStakingRewards;
-import com.hedera.hapi.node.state.token.Nft;
-import com.hedera.hapi.node.state.token.StakingNodeInfo;
-import com.hedera.hapi.node.state.token.Token;
-import com.hedera.hapi.node.state.token.TokenRelation;
-import com.hedera.node.app.service.mono.state.adapters.MerkleMapLike;
-import com.hedera.node.app.service.mono.state.adapters.VirtualMapLike;
 import com.hedera.node.app.service.mono.state.merkle.MerkleNetworkContext;
 import com.hedera.node.app.service.mono.state.merkle.MerkleStakingInfo;
 import com.hedera.node.app.service.mono.state.merkle.MerkleToken;
-import com.hedera.node.app.service.mono.state.merkle.MerkleTokenRelStatus;
-import com.hedera.node.app.service.mono.state.merkle.MerkleUniqueToken;
-import com.hedera.node.app.service.mono.state.migration.AccountStateTranslator;
-import com.hedera.node.app.service.mono.state.migration.NftStateTranslator;
-import com.hedera.node.app.service.mono.state.migration.StakingNodeInfoStateTranslator;
-import com.hedera.node.app.service.mono.state.migration.TokenRelationStateTranslator;
-import com.hedera.node.app.service.mono.state.migration.TokenStateTranslator;
 import com.hedera.node.app.service.mono.state.virtual.EntityNumVirtualKey;
 import com.hedera.node.app.service.mono.state.virtual.UniqueTokenKey;
 import com.hedera.node.app.service.mono.state.virtual.UniqueTokenValue;
@@ -49,21 +30,14 @@ import com.hedera.node.app.service.mono.state.virtual.entities.OnDiskAccount;
 import com.hedera.node.app.service.mono.state.virtual.entities.OnDiskTokenRel;
 import com.hedera.node.app.service.mono.utils.EntityNum;
 import com.hedera.node.app.service.token.TokenService;
-import com.hedera.node.app.service.token.impl.codec.NetworkingStakingTranslator;
 import com.hedera.node.app.service.token.impl.schemas.InitialModServiceTokenSchema;
 import com.hedera.node.app.service.token.impl.schemas.SyntheticRecordsGenerator;
-import com.hedera.node.app.spi.state.MigrationContext;
-import com.hedera.node.app.spi.state.Schema;
 import com.hedera.node.app.spi.state.SchemaRegistry;
-import com.hedera.node.app.spi.state.WritableKVStateBase;
-import com.hedera.node.app.spi.state.WritableSingletonStateBase;
-import com.swirlds.common.threading.manager.AdHocThreadManager;
 import com.swirlds.merkle.map.MerkleMap;
 import com.swirlds.virtualmap.VirtualMap;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Collections;
 import java.util.SortedSet;
-import java.util.function.BiConsumer;
 import java.util.function.Supplier;
 
 /** An implementation of the {@link TokenService} interface. */
@@ -132,157 +106,18 @@ public class TokenServiceImpl implements TokenService {
         requireNonNull(registry);
         // We intentionally ignore the given (i.e. passed-in) version in this method
         registry.register(new InitialModServiceTokenSchema(
-                sysAccts, stakingAccts, treasuryAccts, miscAccts, blocklistAccts, RELEASE_045_VERSION));
-
-        registry.register(new Schema(RELEASE_MIGRATION_VERSION) {
-            @Override
-            public void migrate(@NonNull final MigrationContext ctx) {
-                if (acctsFs != null) {
-                    System.out.println("BBM: migrating token service");
-
-                    // ---------- NFTs
-                    System.out.println("BBM: doing nfts...");
-                    var nftsToState = ctx.newStates().<NftID, Nft>get(NFTS_KEY);
-                    try {
-                        VirtualMapLike.from(nftsFs)
-                                .extractVirtualMapData(
-                                        AdHocThreadManager.getStaticThreadManager(),
-                                        entry -> {
-                                            var nftId = entry.left();
-                                            var toNftId = NftID.newBuilder()
-                                                    .tokenId(TokenID.newBuilder()
-                                                            .tokenNum(nftId.getNum())
-                                                            .build())
-                                                    .serialNumber(nftId.getTokenSerial())
-                                                    .build();
-                                            var fromNft = entry.right();
-                                            var fromNft2 = new MerkleUniqueToken(
-                                                    fromNft.getOwner(),
-                                                    fromNft.getMetadata(),
-                                                    fromNft.getCreationTime());
-                                            var translated = NftStateTranslator.nftFromMerkleUniqueToken(fromNft2);
-                                            nftsToState.put(toNftId, translated);
-                                        },
-                                        1);
-                    } catch (InterruptedException e) {
-                        throw new RuntimeException(e);
-                    }
-                    if (nftsToState.isModified()) ((WritableKVStateBase) nftsToState).commit();
-                    System.out.println("BBM: finished nfts");
-
-                    // ---------- Token Rels/Associations
-                    System.out.println("BBM: doing token rels...");
-                    var tokenRelsToState = ctx.newStates().<EntityIDPair, TokenRelation>get(TOKEN_RELS_KEY);
-                    try {
-                        VirtualMapLike.from(trFs)
-                                .extractVirtualMapData(
-                                        AdHocThreadManager.getStaticThreadManager(),
-                                        entry -> {
-                                            var fromTokenRel = entry.right();
-                                            var key = fromTokenRel.getKey();
-                                            var translated =
-                                                    TokenRelationStateTranslator.tokenRelationFromMerkleTokenRelStatus(
-                                                            new MerkleTokenRelStatus(
-                                                                    fromTokenRel.getBalance(),
-                                                                    fromTokenRel.isFrozen(),
-                                                                    fromTokenRel.isKycGranted(),
-                                                                    fromTokenRel.isAutomaticAssociation(),
-                                                                    fromTokenRel.getNumbers()));
-                                            var newPair = EntityIDPair.newBuilder()
-                                                    .accountId(AccountID.newBuilder()
-                                                            .accountNum(key.getHiOrderAsLong())
-                                                            .build())
-                                                    .tokenId(TokenID.newBuilder()
-                                                            .tokenNum(key.getLowOrderAsLong())
-                                                            .build())
-                                                    .build();
-                                            tokenRelsToState.put(newPair, translated);
-                                        },
-                                        1);
-                    } catch (InterruptedException e) {
-                        throw new RuntimeException(e);
-                    }
-                    if (tokenRelsToState.isModified()) ((WritableKVStateBase) tokenRelsToState).commit();
-                    System.out.println("BBM: finished token rels");
-
-                    // ---------- Accounts
-                    System.out.println("BBM: doing accounts");
-                    var acctsToState = ctx.newStates().<AccountID, Account>get(ACCOUNTS_KEY);
-                    try {
-                        VirtualMapLike.from(acctsFs)
-                                .extractVirtualMapData(
-                                        AdHocThreadManager.getStaticThreadManager(),
-                                        entry -> {
-                                            var acctNum =
-                                                    entry.left().asEntityNum().longValue();
-                                            var fromAcct = entry.right();
-                                            var toAcct = AccountStateTranslator.accountFromOnDiskAccount(fromAcct);
-                                            acctsToState.put(
-                                                    AccountID.newBuilder()
-                                                            .accountNum(acctNum)
-                                                            .build(),
-                                                    toAcct);
-                                        },
-                                        1);
-                    } catch (InterruptedException e) {
-                        throw new RuntimeException(e);
-                    }
-                    if (acctsToState.isModified()) ((WritableKVStateBase) acctsToState).commit();
-                    System.out.println("BBM: finished accts");
-
-                    // ---------- Tokens
-                    System.out.println("BBM: starting tokens (both fung and non-fung)");
-                    var tokensToState = ctx.newStates().<TokenID, Token>get(TOKENS_KEY);
-                    MerkleMapLike.from(tFs).forEachNode(new BiConsumer<EntityNum, MerkleToken>() {
-                        @Override
-                        public void accept(EntityNum entityNum, MerkleToken merkleToken) {
-                            var toToken = TokenStateTranslator.tokenFromMerkle(merkleToken);
-                            tokensToState.put(
-                                    TokenID.newBuilder()
-                                            .tokenNum(entityNum.longValue())
-                                            .build(),
-                                    toToken);
-                        }
-                    });
-                    if (tokensToState.isModified()) ((WritableKVStateBase) tokensToState).commit();
-                    System.out.println("BBM: finished tokens (fung and non-fung)");
-
-                    // ---------- Staking Info
-                    System.out.println("BBM: starting staking info");
-                    var stakingToState = ctx.newStates().<EntityNumber, StakingNodeInfo>get(STAKING_INFO_KEY);
-                    stakingFs.forEach(new BiConsumer<EntityNum, MerkleStakingInfo>() {
-                        @Override
-                        public void accept(EntityNum entityNum, MerkleStakingInfo merkleStakingInfo) {
-                            var toStakingInfo =
-                                    StakingNodeInfoStateTranslator.stakingInfoFromMerkleStakingInfo(merkleStakingInfo);
-                            stakingToState.put(
-                                    EntityNumber.newBuilder()
-                                            .number(merkleStakingInfo.getKey().longValue())
-                                            .build(),
-                                    toStakingInfo);
-                        }
-                    });
-                    if (stakingToState.isModified()) ((WritableKVStateBase) stakingToState).commit();
-                    System.out.println("BBM: finished staking info");
-
-                    // ---------- Staking Rewards
-                    System.out.println("BBM: starting staking rewards");
-                    var srToState = ctx.newStates().<NetworkStakingRewards>getSingleton(STAKING_NETWORK_REWARDS_KEY);
-                    var toSr = NetworkingStakingTranslator.networkStakingRewardsFromMerkleNetworkContext(mnc);
-                    srToState.put(toSr);
-                    if (srToState.isModified()) ((WritableSingletonStateBase) srToState).commit();
-                    System.out.println("BBM: finished staking rewards");
-
-                    nftsFs = null;
-                    trFs = null;
-                    acctsFs = null;
-                    tFs = null;
-
-                    stakingFs = null;
-                    mnc = null;
-                }
-            }
-        });
+                sysAccts,
+                stakingAccts,
+                treasuryAccts,
+                miscAccts,
+                blocklistAccts,
+                version,
+                acctsFs,
+                tFs,
+                stakingFs,
+                nftsFs,
+                trFs,
+                mnc));
     }
 
     public void setNftsFromState(VirtualMap<UniqueTokenKey, UniqueTokenValue> fs) {
