+++ hedera-node/hedera-token-service-impl/src/main/java/com/hedera/node/app/service/token/impl/schemas/InitialModServiceTokenSchema.java
@@ -41,20 +41,45 @@ import com.hedera.hapi.node.state.token.Nft;
 import com.hedera.hapi.node.state.token.StakingNodeInfo;
 import com.hedera.hapi.node.state.token.Token;
 import com.hedera.hapi.node.state.token.TokenRelation;
+import com.hedera.node.app.service.mono.state.adapters.MerkleMapLike;
+import com.hedera.node.app.service.mono.state.adapters.VirtualMapLike;
+import com.hedera.node.app.service.mono.state.merkle.MerkleNetworkContext;
+import com.hedera.node.app.service.mono.state.merkle.MerkleStakingInfo;
+import com.hedera.node.app.service.mono.state.merkle.MerkleToken;
+import com.hedera.node.app.service.mono.state.merkle.MerkleTokenRelStatus;
+import com.hedera.node.app.service.mono.state.merkle.MerkleUniqueToken;
+import com.hedera.node.app.service.mono.state.migration.AccountStateTranslator;
+import com.hedera.node.app.service.mono.state.migration.NftStateTranslator;
+import com.hedera.node.app.service.mono.state.migration.StakingNodeInfoStateTranslator;
+import com.hedera.node.app.service.mono.state.migration.TokenRelationStateTranslator;
+import com.hedera.node.app.service.mono.state.migration.TokenStateTranslator;
+import com.hedera.node.app.service.mono.state.virtual.EntityNumVirtualKey;
+import com.hedera.node.app.service.mono.state.virtual.UniqueTokenKey;
+import com.hedera.node.app.service.mono.state.virtual.UniqueTokenValue;
+import com.hedera.node.app.service.mono.state.virtual.entities.OnDiskAccount;
+import com.hedera.node.app.service.mono.state.virtual.entities.OnDiskTokenRel;
+import com.hedera.node.app.service.mono.utils.EntityNum;
 import com.hedera.node.app.service.token.impl.TokenServiceImpl;
+import com.hedera.node.app.service.token.impl.codec.NetworkingStakingTranslator;
 import com.hedera.node.app.spi.state.MigrationContext;
 import com.hedera.node.app.spi.state.Schema;
 import com.hedera.node.app.spi.state.StateDefinition;
+import com.hedera.node.app.spi.state.WritableKVStateBase;
+import com.hedera.node.app.spi.state.WritableSingletonStateBase;
 import com.hedera.node.config.data.AccountsConfig;
 import com.hedera.node.config.data.HederaConfig;
 import com.hedera.node.config.data.LedgerConfig;
 import com.hedera.node.config.data.StakingConfig;
 import com.hedera.pbj.runtime.io.buffer.Bytes;
+import com.swirlds.common.threading.manager.AdHocThreadManager;
+import com.swirlds.merkle.map.MerkleMap;
+import com.swirlds.virtualmap.VirtualMap;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Arrays;
 import java.util.Set;
 import java.util.SortedSet;
 import java.util.TreeSet;
+import java.util.function.BiConsumer;
 import java.util.function.Supplier;
 import java.util.stream.LongStream;
 import org.apache.logging.log4j.LogManager;
@@ -84,6 +109,13 @@ public class InitialModServiceTokenSchema extends Schema {
     private final Supplier<SortedSet<Account>> miscAccts;
     private final Supplier<SortedSet<Account>> blocklistAccts;
 
+    private VirtualMap<EntityNumVirtualKey, OnDiskAccount> acctsFs;
+    private MerkleMap<EntityNum, MerkleToken> tFs;
+    private MerkleMap<EntityNum, MerkleStakingInfo> stakingFs;
+    private VirtualMap<UniqueTokenKey, UniqueTokenValue> nftsFs;
+    private VirtualMap<EntityNumVirtualKey, OnDiskTokenRel> trFs;
+    private MerkleNetworkContext mnc;
+
     /**
      * Constructor for this schema. Each of the supplier params should produce a {@link SortedSet} of
      * {@link Account} objects, where each account object represents a _synthetic record_ (see {@link
@@ -92,19 +124,31 @@ public class InitialModServiceTokenSchema extends Schema {
      * but are required for an event recovery situation.
      */
     public InitialModServiceTokenSchema(
-            @NonNull final Supplier<SortedSet<Account>> sysAcctRcds,
-            @NonNull final Supplier<SortedSet<Account>> stakingAcctRcds,
-            @NonNull final Supplier<SortedSet<Account>> treasuryAcctRcds,
-            @NonNull final Supplier<SortedSet<Account>> miscAcctRcds,
-            @NonNull final Supplier<SortedSet<Account>> blocklistAcctRcds,
-            @NonNull final SemanticVersion version) {
+            final Supplier<SortedSet<Account>> sysAccts,
+            final Supplier<SortedSet<Account>> stakingAccts,
+            final Supplier<SortedSet<Account>> treasuryAccts,
+            final Supplier<SortedSet<Account>> miscAccts,
+            final Supplier<SortedSet<Account>> blocklistAccts,
+            final SemanticVersion version,
+            final VirtualMap<EntityNumVirtualKey, OnDiskAccount> acctsFs,
+            final MerkleMap<EntityNum, MerkleToken> tFs,
+            final MerkleMap<EntityNum, MerkleStakingInfo> stakingFs,
+            final VirtualMap<UniqueTokenKey, UniqueTokenValue> nftsFs,
+            final VirtualMap<EntityNumVirtualKey, OnDiskTokenRel> trFs,
+            final MerkleNetworkContext mnc) {
         super(version);
 
-        this.sysAccts = sysAcctRcds;
-        this.stakingAccts = stakingAcctRcds;
-        this.treasuryAccts = treasuryAcctRcds;
-        this.miscAccts = miscAcctRcds;
-        this.blocklistAccts = blocklistAcctRcds;
+        this.sysAccts = sysAccts;
+        this.stakingAccts = stakingAccts;
+        this.treasuryAccts = treasuryAccts;
+        this.miscAccts = miscAccts;
+        this.blocklistAccts = blocklistAccts;
+        this.acctsFs = acctsFs;
+        this.tFs = tFs;
+        this.stakingFs = stakingFs;
+        this.nftsFs = nftsFs;
+        this.trFs = trFs;
+        this.mnc = mnc;
     }
 
     @NonNull
@@ -125,6 +169,142 @@ public class InitialModServiceTokenSchema extends Schema {
         final var isGenesis = ctx.previousStates().isEmpty();
         if (isGenesis) {
             createGenesisSchema(ctx);
+        } else if (acctsFs != null) {
+            System.out.println("BBM: migrating token service");
+
+            // ---------- NFTs
+            System.out.println("BBM: doing nfts...");
+            var nftsToState = ctx.newStates().<NftID, Nft>get(NFTS_KEY);
+            try {
+                VirtualMapLike.from(nftsFs)
+                        .extractVirtualMapData(
+                                AdHocThreadManager.getStaticThreadManager(),
+                                entry -> {
+                                    var nftId = entry.left();
+                                    var toNftId = NftID.newBuilder()
+                                            .tokenId(TokenID.newBuilder()
+                                                    .tokenNum(nftId.getNum())
+                                                    .build())
+                                            .serialNumber(nftId.getTokenSerial())
+                                            .build();
+                                    var fromNft = entry.right();
+                                    var fromNft2 = new MerkleUniqueToken(
+                                            fromNft.getOwner(), fromNft.getMetadata(), fromNft.getCreationTime());
+                                    var translated = NftStateTranslator.nftFromMerkleUniqueToken(fromNft2);
+                                    nftsToState.put(toNftId, translated);
+                                },
+                                1);
+            } catch (InterruptedException e) {
+                throw new RuntimeException(e);
+            }
+            if (nftsToState.isModified()) ((WritableKVStateBase) nftsToState).commit();
+            System.out.println("BBM: finished nfts");
+
+            // ---------- Token Rels/Associations
+            System.out.println("BBM: doing token rels...");
+            var tokenRelsToState = ctx.newStates().<EntityIDPair, TokenRelation>get(TOKEN_RELS_KEY);
+            try {
+                VirtualMapLike.from(trFs)
+                        .extractVirtualMapData(
+                                AdHocThreadManager.getStaticThreadManager(),
+                                entry -> {
+                                    var fromTokenRel = entry.right();
+                                    var key = fromTokenRel.getKey();
+                                    var translated = TokenRelationStateTranslator.tokenRelationFromMerkleTokenRelStatus(
+                                            new MerkleTokenRelStatus(
+                                                    fromTokenRel.getBalance(),
+                                                    fromTokenRel.isFrozen(),
+                                                    fromTokenRel.isKycGranted(),
+                                                    fromTokenRel.isAutomaticAssociation(),
+                                                    fromTokenRel.getNumbers()));
+                                    var newPair = EntityIDPair.newBuilder()
+                                            .accountId(AccountID.newBuilder()
+                                                    .accountNum(key.getHiOrderAsLong())
+                                                    .build())
+                                            .tokenId(TokenID.newBuilder()
+                                                    .tokenNum(key.getLowOrderAsLong())
+                                                    .build())
+                                            .build();
+                                    tokenRelsToState.put(newPair, translated);
+                                },
+                                1);
+            } catch (InterruptedException e) {
+                throw new RuntimeException(e);
+            }
+            if (tokenRelsToState.isModified()) ((WritableKVStateBase) tokenRelsToState).commit();
+            System.out.println("BBM: finished token rels");
+
+            // ---------- Accounts
+            System.out.println("BBM: doing accounts");
+            var acctsToState = ctx.newStates().<AccountID, Account>get(ACCOUNTS_KEY);
+            try {
+                VirtualMapLike.from(acctsFs)
+                        .extractVirtualMapData(
+                                AdHocThreadManager.getStaticThreadManager(),
+                                entry -> {
+                                    var acctNum = entry.left().asEntityNum().longValue();
+                                    var fromAcct = entry.right();
+                                    var toAcct = AccountStateTranslator.accountFromOnDiskAccount(fromAcct);
+                                    acctsToState.put(
+                                            AccountID.newBuilder()
+                                                    .accountNum(acctNum)
+                                                    .build(),
+                                            toAcct);
+                                },
+                                1);
+            } catch (InterruptedException e) {
+                throw new RuntimeException(e);
+            }
+            if (acctsToState.isModified()) ((WritableKVStateBase) acctsToState).commit();
+            System.out.println("BBM: finished accts");
+
+            // ---------- Tokens
+            System.out.println("BBM: starting tokens (both fung and non-fung)");
+            var tokensToState = ctx.newStates().<TokenID, Token>get(TOKENS_KEY);
+            MerkleMapLike.from(tFs).forEachNode(new BiConsumer<EntityNum, MerkleToken>() {
+                @Override
+                public void accept(EntityNum entityNum, MerkleToken merkleToken) {
+                    var toToken = TokenStateTranslator.tokenFromMerkle(merkleToken);
+                    tokensToState.put(
+                            TokenID.newBuilder().tokenNum(entityNum.longValue()).build(), toToken);
+                }
+            });
+            if (tokensToState.isModified()) ((WritableKVStateBase) tokensToState).commit();
+            System.out.println("BBM: finished tokens (fung and non-fung)");
+
+            // ---------- Staking Info
+            System.out.println("BBM: starting staking info");
+            var stakingToState = ctx.newStates().<EntityNumber, StakingNodeInfo>get(STAKING_INFO_KEY);
+            stakingFs.forEach(new BiConsumer<EntityNum, MerkleStakingInfo>() {
+                @Override
+                public void accept(EntityNum entityNum, MerkleStakingInfo merkleStakingInfo) {
+                    var toStakingInfo =
+                            StakingNodeInfoStateTranslator.stakingInfoFromMerkleStakingInfo(merkleStakingInfo);
+                    stakingToState.put(
+                            EntityNumber.newBuilder()
+                                    .number(merkleStakingInfo.getKey().longValue())
+                                    .build(),
+                            toStakingInfo);
+                }
+            });
+            if (stakingToState.isModified()) ((WritableKVStateBase) stakingToState).commit();
+            System.out.println("BBM: finished staking info");
+
+            // ---------- Staking Rewards
+            System.out.println("BBM: starting staking rewards");
+            var srToState = ctx.newStates().<NetworkStakingRewards>getSingleton(STAKING_NETWORK_REWARDS_KEY);
+            var toSr = NetworkingStakingTranslator.networkStakingRewardsFromMerkleNetworkContext(mnc);
+            srToState.put(toSr);
+            if (srToState.isModified()) ((WritableSingletonStateBase) srToState).commit();
+            System.out.println("BBM: finished staking rewards");
+
+            nftsFs = null;
+            trFs = null;
+            acctsFs = null;
+            tFs = null;
+
+            stakingFs = null;
+            mnc = null;
         }
     }
